refactor: keep the api reference settings in their own module
The module held two things: the settings read from a config block, and
the facade that builds a reference from them. The resolver needs only the
former, and now depends on only that.
